Move app-paths
